pyproject.toml
[tool.poetry]
name = "vault-dashboard-streamlit-poc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.8.1"
boto3 = "^1.21.46"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
